+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.34.0
+
+- New feature: for Adam, provide mode = off, related to the regulation_mode = off, and mode = cool, for regulation_mode = cooling.
+  Also, for Adam, return to providing a single setpoint for both heating and cooling.
+- Update userdata, manual-fixture-creation, and tests related to the added modes.
+- A few small Typing updates
+- Manually change the adam_jip fixture, for testing in pw-beta/Core plugwise
+
 ## v0.33.2 Bugfix for HA climate
 
 - Remove last_used (schedule) from output.